+++ src/base/state/abstract_state.jl
@@ -70,6 +70,18 @@ THe default returns `agst.X`.
 get_gradient(agst::AbstractGradientSolverState) = agst.X
 
 """
+    get_gradient(s::AbstractManoptSolverState)
+
+return the (last stored) gradient within [`AbstractManoptSolverState`](@ref)` `s`.
+By default also undecorates the state beforehand
+"""
+get_gradient(s::AbstractManoptSolverState) = _get_gradient(s, dispatch_state_decorator(s))
+function _get_gradient(s::AbstractManoptSolverState, ::Val{false})
+    return error("It seems that $s do not provide access to a gradient")
+end
+_get_gradient(s::AbstractManoptSolverState, ::Val{true}) = get_gradient(s.state)
+
+"""
     get_iterate(state::AbstractManoptSolverState)
 
 return the (last stored) iterate within [`AbstractManoptSolverState`](@ref)` `state`.
@@ -95,9 +107,6 @@ THe default returns `agst.p`.
 """
 get_iterate(agst::AbstractGradientSolverState) = agst.p
 
-
-_set_iterate!(s::AbstractManoptSolverState, M, p, ::Val{true}) = set_iterate!(s.state, M, p)
-
 @doc """
     get_message(du::AbstractManoptSolverState)
 
@@ -110,6 +119,86 @@ end
 _get_message(s::AbstractManoptSolverState, ::Val{true}) = get_message(s.state)
 #INtroduce a default that there is no message
 _get_message(s::AbstractManoptSolverState, ::Val{false}) = ""
+
+"""
+    get_solver_return(s::AbstractManoptSolverState)
+    get_solver_return(o::AbstractManifoldObjective, s::AbstractManoptSolverState)
+
+determine the result value of a call to a solver.
+By default this returns the same as [`get_solver_result`](@ref).
+
+    get_solver_return(o::ReturnManifoldObjective, s::AbstractManoptSolverState)
+
+return both the objective and the state as a tuple.
+"""
+function get_solver_return(s::AbstractManoptSolverState)
+    return _get_solver_return(s, dispatch_state_decorator(s))
+end
+_get_solver_return(s::AbstractManoptSolverState, ::Val{false}) = get_solver_result(s)
+_get_solver_return(s::AbstractManoptSolverState, ::Val{true}) = get_solver_return(s.state)
+
+# also work in combination with the objective
+function get_solver_return(o::AbstractManifoldObjective, s::AbstractManoptSolverState)
+    #resolve objective first
+    return _get_solver_return(o, s, dispatch_objective_decorator(o))
+end
+# remove decorator
+function _get_solver_return(o::AbstractManifoldObjective, s, ::Val{true})
+    return get_solver_return(get_objective(o, false), s)
+end
+_get_solver_return(::AbstractManifoldObjective, s, ::Val{false}) = get_solver_return(s)
+function get_solver_return(o::ReturnManifoldObjective, s::AbstractManoptSolverState)
+    return o.objective, get_solver_return(s)
+end
+
+"""
+    get_solver_result(state::AbstractManoptSolverState)
+    get_solver_result(tos::Tuple{AbstractManifoldObjective,AbstractManoptSolverState})
+    get_solver_result(objective::AbstractManifoldObjective, state::AbstractManoptSolverState)
+    get_solver_result(problem::AbstractManoptProblem, state::AbstractManoptSolverState)
+
+Return the final result after all iterations that is stored within
+the [`AbstractManoptSolverState`](@ref) `ams`, which was modified during the iterations.
+
+For the case an [`AbstractManifoldObjective`](@ref) `o` the objective is passed as well
+– either as a Tuple or as two parameters –, by default, the objective is ignored,
+and the solver result for the state is called; this is due to display reasons in REPL
+related to statistics, where such a Tuple might appear
+
+For the case an [`AbstractManoptProblem`](@ref) `p` is passed as well as
+a first optional parameter, by default the problem is ignored.
+This can be used to change the representation of a result stored in a state, e.g.
+when a tangent vector is (part of) the result, changing between representations in
+coefficients and different tangent vector representations could be performed as a final step,
+depending on which problem was aimed to be solved
+
+Note that the returned value or point might still be aliased to the original `state`.
+"""
+function get_solver_result(state::AbstractManoptSolverState)
+    return _get_solver_result(state, dispatch_state_decorator(state))
+end
+function get_solver_result(
+        tos::Tuple{<:AbstractManifoldObjective, <:AbstractManoptSolverState}
+    )
+    return get_solver_result(tos...)
+end
+function get_solver_result(::AbstractManifoldObjective, state::AbstractManoptSolverState)
+    return get_solver_result(state)
+end
+#A problem or – hence untyped – a closed form solution / function
+function get_solver_result(pf, state::AbstractManoptSolverState)
+    return get_solver_result(state)
+end
+function get_solver_result(tos::Tuple{<:AbstractManifoldObjective, S}) where {S}
+    return tos[2]
+end
+# if the second one is anything else, assume it is a point/result -> return that
+function get_solver_result(::AbstractManifoldObjective, p)
+    return p
+end
+_get_solver_result(state::AbstractManoptSolverState, ::Val{false}) = get_iterate(state)
+_get_solver_result(state::AbstractManoptSolverState, ::Val{true}) = get_solver_result(state.state)
+
 
 @doc """
     get_state(s::AbstractManoptSolverState, recursive::Bool=true)
@@ -166,6 +255,24 @@ function set_gradient!(state::AbstractGradientSolverState, M, p, X)
 end
 
 """
+    set_gradient!(s::AbstractManoptSolverState, M::AbstractManifold, p, X)
+
+set the gradient within an (possibly decorated) [`AbstractManoptSolverState`](@ref)
+to some (start) value `X` in the tangent space at `p`.
+"""
+function set_gradient!(s::AbstractManoptSolverState, M, p, X)
+    return _set_gradient!(s, M, p, X, dispatch_state_decorator(s))
+end
+function _set_gradient!(s::AbstractManoptSolverState, ::Any, ::Any, ::Any, ::Val{false})
+    return error(
+        "It seems the AbstractManoptSolverState $s do not provide (write) access to a gradient",
+    )
+end
+function _set_gradient!(s::AbstractManoptSolverState, M, p, X, ::Val{true})
+    return set_gradient!(s.state, M, p, X)
+end
+
+"""
     set_iterate!(s::AbstractManoptSolverState, M::AbstractManifold, p)
 
 set the iterate within an [`AbstractManoptSolverState`](@ref) to some (start) value `p`.
@@ -178,6 +285,7 @@ function _set_iterate!(s::AbstractManoptSolverState, ::Any, ::Any, ::Val{false})
         "It seems the AbstractManoptSolverState $s do not provide (write) access to an iterate",
     )
 end
+_set_iterate!(s::AbstractManoptSolverState, M, p, ::Val{true}) = set_iterate!(s.state, M, p)
 
 """
     set_parameter!(ams::AbstractManoptSolverState, element::Symbol, args...)